requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
email-validator==2.1.0
psycopg2-binary==2.9.9
pytest==7.4.3
Werkzeug==2.3.7